# docs/indexer/indexer_swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Meta Media Indexer Service API, provides file query and download functionality",
        "title": "Meta Media Indexer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:7281",
    "basePath": "/api/v1",
    "paths": {
        "/files": {
            "get": {
                "description": "Query file list with cursor pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indexer File Query"
                ],
                "summary": "Query file list",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Cursor (last file ID)",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-media-service_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "files": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/definitions/model.IndexerFile"
                                                    }
                                                },
                                                "has_more": {
                                                    "type": "boolean"
                                                },
                                                "next_cursor": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta-media-service_controller_respond.Response"
                        }
                    }
                }
            }
        },
        "/files/creator/{address}": {
            "get": {
                "description": "Query file list by creator address with cursor pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indexer File Query"
                ],
                "summary": "Get files by creator address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Creator address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Cursor (last file ID)",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-media-service_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "files": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/definitions/model.IndexerFile"
                                                    }
                                                },
                                                "has_more": {
                                                    "type": "boolean"
                                                },
                                                "next_cursor": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta-media-service_controller_respond.Response"
                        }
                    }
                }
            }
        },
        "/files/metaid/{metaId}": {
            "get": {
                "description": "Query file list by creator MetaID with cursor pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indexer File Query"
                ],
                "summary": "Get files by creator MetaID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Creator MetaID",
                        "name": "metaId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Cursor (last file ID)",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-media-service_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "files": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/definitions/model.IndexerFile"
                                                    }
                                                },
                                                "has_more": {
                                                    "type": "boolean"
                                                },
                                                "next_cursor": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta-media-service_controller_respond.Response"
                        }
                    }
                }
            }
        },
        "/files/{pinId}": {
            "get": {
                "description": "Query file details by PIN ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indexer File Query"
                ],
                "summary": "Get file by PIN ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PIN ID",
                        "name": "pinId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-media-service_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.IndexerFile"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/meta-media-service_controller_respond.Response"
                        }
                    }
                }
            }
        },
        "/files/{pinId}/content": {
            "get": {
                "description": "Get file content by PIN ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Indexer File Query"
                ],
                "summary": "Get file content",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PIN ID",
                        "name": "pinId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/meta-media-service_controller_respond.Response"
                        }
                    }
                }
            }
        },
        "/stats": {
            "get": {
                "description": "Get indexer statistics (total files count, etc.)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indexer Status"
                ],
                "summary": "Get statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-media-service_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta-media-service_controller_respond.Response"
                        }
                    }
                }
            }
        },
        "/status": {
            "get": {
                "description": "Get indexer synchronization status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indexer Status"
                ],
                "summary": "Get sync status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/meta-media-service_controller_respond.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.IndexerSyncStatus"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/meta-media-service_controller_respond.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "meta-media-service_controller_respond.Response": {
            "description": "Unified API response structure",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "processingTime": {
                    "type": "integer",
                    "example": 123
                }
            }
        },
        "model.IndexerFile": {
            "type": "object",
            "properties": {
                "block_height": {
                    "description": "Block height",
                    "type": "integer"
                },
                "chain_name": {
                    "description": "Blockchain related fields",
                    "type": "string"
                },
                "content_type": {
                    "description": "Content type",
                    "type": "string"
                },
                "created_at": {
                    "description": "Timestamps",
                    "type": "string"
                },
                "creator_address": {
                    "description": "Creator address",
                    "type": "string"
                },
                "creator_meta_id": {
                    "description": "Creator MetaID (SHA256 hash)",
                    "type": "string"
                },
                "encryption": {
                    "description": "Encryption method",
                    "type": "string"
                },
                "file_md5": {
                    "description": "File MD5",
                    "type": "string"
                },
                "file_name": {
                    "description": "File related fields",
                    "type": "string"
                },
                "file_size": {
                    "description": "File size",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "operation": {
                    "description": "create/modify/revoke",
                    "type": "string"
                },
                "parent_path": {
                    "description": "Parent path",
                    "type": "string"
                },
                "path": {
                    "description": "MetaID path",
                    "type": "string"
                },
                "pin_id": {
                    "description": "MetaID related fields",
                    "type": "string"
                },
                "state": {
                    "description": "State 0:EXIST,2:DELETED",
                    "type": "integer"
                },
                "status": {
                    "description": "Status fields",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Status"
                        }
                    ]
                },
                "storage_path": {
                    "description": "Storage path",
                    "type": "string"
                },
                "storage_type": {
                    "description": "Storage related fields",
                    "type": "string"
                },
                "timestamp": {
                    "description": "Timestamp (seconds since epoch)",
                    "type": "integer"
                },
                "tx_id": {
                    "description": "Transaction ID",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Update time",
                    "type": "string"
                },
                "version": {
                    "description": "Version",
                    "type": "string"
                },
                "vout": {
                    "description": "Output index",
                    "type": "integer"
                }
            }
        },
        "model.IndexerSyncStatus": {
            "type": "object",
            "properties": {
                "chain_name": {
                    "description": "Chain information",
                    "type": "string"
                },
                "created_at": {
                    "description": "Timestamps",
                    "type": "string"
                },
                "current_sync_height": {
                    "description": "Sync status",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "description": "Update time",
                    "type": "string"
                }
            }
        },
        "model.Status": {
            "type": "string",
            "enum": [
                "pending",
                "success",
                "failed"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusSuccess",
                "StatusFailed"
            ]
        }
    }
}